+++ src/comment.ts
@@ -2,6 +2,9 @@ import { Context } from '@actions/github/lib/context';
 import { GitHub } from '@actions/github/lib/utils';
 import { PullRequestEvent } from '@octokit/webhooks-definitions/schema';
 
+/**
+ * Comments controls interacting with GitHub to make comments
+ */
 export class Comments {
   private readonly issueNumber: number;
   private readonly commitSha: string;
@@ -17,6 +20,12 @@ export class Comments {
     this.issueNumber = this.context.payload.pull_request?.number!;
   }
 
+  /**
+   * Find the previous comment with the given hash
+   *
+   * @param hash the unique hash identifying the stage comment to look for
+   * @returns the PR comment id or undefined if there is no previous comment
+   */
   public async findPrevious(hash: string): Promise<number | undefined> {
     const comments = await this.octokit.rest.issues.listComments({
       ...this.context.repo,
@@ -25,6 +34,13 @@ export class Comments {
     return comments.data.find(comment => comment.body?.includes(hash))?.id;
   }
 
+  /**
+   * Update an existing comment
+   *
+   * @param hash the unique hash identifying the stage comment to look for
+   * @param content the content of the comment
+   * @param commentId the id of the comment to update
+   */
   public async updateComment(commentId: number, hash: string, content: string[]) {
     await this.octokit.rest.issues.updateComment({
       ...this.context.repo,
@@ -38,6 +54,12 @@ export class Comments {
     });
   }
 
+  /**
+   * Create a new comment
+   *
+   * @param hash the unique hash identifying the stage comment to look for
+   * @param content the content of the comment
+   */
   public async createComment(hash: string, content: string[]) {
     await this.octokit.rest.issues.createComment({
       ...this.context.repo,
@@ -50,5 +72,4 @@ export class Comments {
       issue_number: this.issueNumber,
     });
   }
-
 }
