+++ test/diff.test.ts
@@ -2,7 +2,7 @@ import { ResourceImpact } from '@aws-cdk/cloudformation-diff';
 import { CloudFormationClient, GetTemplateCommand } from '@aws-sdk/client-cloudformation';
 import { mockClient } from 'aws-sdk-client-mock';
 import { StackInfo } from '../src/assembly';
-import { StackDiff, StageProcessor } from '../src/diff';
+import { StackDiff } from '../src/diff';
 
 const cfnMock = mockClient(CloudFormationClient);
 
@@ -302,72 +302,4 @@ describe('StackDiff', () => {
       destructiveChanges: [],
     });
   });
-});
-
-describe('StageProcessor', () => {
-  const stackInfo: StackInfo = {
-    name: 'my-stack',
-    content: {
-      Resources: {
-        MyRole: {
-          Type: 'AWS::IAM::Role',
-          Properties: {
-            RoleName: 'MyCustomName',
-          },
-        },
-      },
-    },
-  };
-
-  test('stage with no diffs', async () => {
-    cfnMock.on(GetTemplateCommand)
-      .resolves({
-        TemplateBody: JSON.stringify(stackInfo.content),
-      });
-    const processor = new StageProcessor([
-      {
-        name: 'Stage1',
-        stacks: [stackInfo],
-      },
-    ], []);
-    await processor.processStages();
-    const p = (processor as any).stageComments;
-    expect(p).toEqual({
-      Stage1: expect.any(Object),
-    });
-    expect(p.Stage1.comment).toEqual(['No Changes for stack: my-stack']);
-  });
-
-  test('stage with diff', async () => {
-    cfnMock.on(GetTemplateCommand)
-      .resolves({
-        TemplateBody: JSON.stringify(stackInfo.content),
-      });
-    const processor = new StageProcessor([
-      {
-        name: 'Stage1',
-        stacks: [{
-          name: 'my-stack',
-          content: {
-            Resources: {
-              MyRole: {
-                Type: 'AWS::IAM::Role',
-                Properties: {
-                  RoleName: 'MyNewCustomName',
-                },
-              },
-            },
-          },
-        }],
-      },
-    ], []);
-    await processor.processStages();
-    const p = (processor as any).stageComments;
-    expect(p).toEqual({
-      Stage1: expect.any(Object),
-    });
-    expect(p.Stage1.comment.length).not.toEqual(0);
-    expect(p.Stage1.destructiveChanges).toEqual(1);
-  });
-
 });
